Move ConcertProjection into example.adapter.projection package
