+++ sample-app/src/main/scala/example/adapter/projection/ConcertProjection.scala
@@ -1,4 +1,4 @@
-package example.adapter
+package example.adapter.projection
 
 trait ConcertProjection {
   def start(): Unit
